+++ mock-api-test-sdk-net80/users/TestUserDeactivate.cs
@@ -1,0 +1,60 @@
+using System.Web;
+using Newtonsoft.Json;
+using Smartsheet.Api;
+using Smartsheet.Api.Models;
+
+namespace mock_api_test_sdk_net80
+{
+    [TestClass]
+    public class TestUserDeactivate
+    {
+        [TestMethod]
+        public async Task TestDeactivateUserGeneratedUrlIsCorrect()
+        {
+            Guid requestId = Guid.NewGuid();
+            SmartsheetClient ss = HelperFunctions.SetupClient("/users/deactivate-user/all-response-body-properties", requestId.ToString());
+
+            ss.UserResources.DeactivateUser(CommonTestConstants.TEST_USER_ID);
+            WiremockHelper wiremockHelper = new WiremockHelper();
+            LogModel foundRequest = await wiremockHelper.FindWiremockRequestAsync(requestId.ToString());
+            var uri = new Uri(foundRequest.AbsoluteUrl);
+            string path = uri.AbsolutePath;
+
+            Assert.AreEqual($"/2.0/users/{CommonTestConstants.TEST_USER_ID}/deactivate", path);
+        }
+
+        [TestMethod]
+        public async Task TestDeactivateUserAllResponseBodyProperties()
+        {
+            Guid requestId = Guid.NewGuid();
+            SmartsheetClient ss = HelperFunctions.SetupClient("/users/deactivate-user/all-response-body-properties", requestId.ToString());
+
+            // If this throws an exception, the test will fail automatically
+            ss.UserResources.DeactivateUser(CommonTestConstants.TEST_USER_ID);
+        }
+
+        [TestMethod]
+        public void TestDeactivateUserError500Response()
+        {
+            Guid requestId = Guid.NewGuid();
+            SmartsheetClient ss = HelperFunctions.SetupClient("/errors/500-response", requestId.ToString());
+
+            SmartsheetException exception = Assert.ThrowsException<SmartsheetException>(() =>
+                ss.UserResources.DeactivateUser(CommonTestConstants.TEST_USER_ID));
+            
+            Assert.IsTrue(exception.Message.Contains("Internal Server Error"));
+        }
+
+        [TestMethod]
+        public void TestDeactivateUserError400Response()
+        {
+            Guid requestId = Guid.NewGuid();
+            SmartsheetClient ss = HelperFunctions.SetupClient("/errors/400-response", requestId.ToString());
+
+            SmartsheetException exception = Assert.ThrowsException<SmartsheetException>(() =>
+                ss.UserResources.DeactivateUser(CommonTestConstants.TEST_USER_ID));
+            
+            Assert.IsTrue(exception.Message.Contains("Malformed Request"));
+        }
+    }
+}